gnd: Share ABI preprocessing between codegen and validation
Move preprocess_abi_json (and add_default_event_param_names) out of the
codegen command and into the shared abi module so that other components
can normalize ABIs the same way codegen does. No behavior change for
codegen; this is a prerequisite for making manifest validation parse
ABIs identically to codegen.
