+++ gnd/src/abi.rs
@@ -40,6 +40,62 @@ pub fn normalize_abi_json(abi_str: &str) -> Result<serde_json::Value> {
     ))
 }
 
+/// Preprocess ABI JSON to normalize artifact formats and add defaults
+/// required by alloy's ABI parser:
+/// - `anonymous: false` for events (alloy requires this field)
+/// - `param{index}` names for unnamed event parameters (to match graph-cli behavior)
+///
+/// This is shared by codegen and validation so that both parse the ABI in
+/// exactly the same way: an ABI that codegen accepts must also be accepted by
+/// validation (and vice versa).
+pub fn preprocess_abi_json(abi_str: &str) -> Result<String> {
+    // Normalize to get the ABI array from various artifact formats
+    let mut abi = normalize_abi_json(abi_str)?;
+
+    if let Some(items) = abi.as_array_mut() {
+        for item in items {
+            if let Some(obj) = item.as_object_mut() {
+                let is_event = obj
+                    .get("type")
+                    .and_then(|t| t.as_str())
+                    .map(|t| t == "event")
+                    .unwrap_or(false);
+
+                if is_event {
+                    // Add anonymous: false for events if missing (alloy requires it)
+                    if !obj.contains_key("anonymous") {
+                        obj.insert("anonymous".to_string(), serde_json::Value::Bool(false));
+                    }
+
+                    // Add param{index} names for unnamed event parameters
+                    if let Some(inputs) = obj.get_mut("inputs") {
+                        add_default_event_param_names(inputs);
+                    }
+                }
+            }
+        }
+    }
+
+    serde_json::to_string(&abi).context("Failed to serialize processed ABI")
+}
+
+/// Add `param{index}` names to unnamed event parameters to match graph-cli behavior.
+/// Alloy defaults missing names to empty strings, but for events we want `param0`, `param1`, etc.
+fn add_default_event_param_names(params: &mut serde_json::Value) {
+    if let Some(params_arr) = params.as_array_mut() {
+        for (index, param) in params_arr.iter_mut().enumerate() {
+            if let Some(obj) = param.as_object_mut()
+                && !obj.contains_key("name")
+            {
+                obj.insert(
+                    "name".to_string(),
+                    serde_json::Value::String(format!("param{}", index)),
+                );
+            }
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
